# Uritmix.Api/wwwroot/swagger/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Uritmix.Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Авторизация пользователя",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUser"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUser"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUser"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUser"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidError"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLoggedUser"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Получение нового token с помощью refresh token",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Refresh"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Refresh"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Refresh"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Refresh"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidError"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLoggedUser"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/password-reset-query": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Запрос на смену пароля",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetQuery"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetQuery"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidError"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUnit"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Смена пароля на основе токена отправленного на почту",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidError"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoggedUser": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRoleView"
          },
          "email": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "LoginUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PasswordReset": {
        "type": "object",
        "properties": {
          "confirmCode": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "passwordConfirm": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PasswordResetQuery": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PropertyError": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Refresh": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ResultBoolean": {
        "type": "object",
        "properties": {
          "result": {
            "type": "boolean"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ResultLoggedUser": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/LoggedUser"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ResultUnit": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/Unit"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Unit": {
        "type": "object",
        "additionalProperties": false
      },
...
      "UserRoleView": {
        "enum": [
          "User",
          "Admin"
        ],
        "type": "string"
      },
...
      "ValidError": {
        "type": "object",
        "properties": {
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyError"
            }
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}